Updated controllers with dashboardRoutes.js, and relocated GET routes from askRoutes/giveRoutes.js to homeRoutes.js
