+++ assets/psuedocode
@@ -8,9 +8,10 @@ Ask_Give
 Resources
 Logout
 
-//Landing
+//Landing [Page: ]
 Navigate to the application url
-Present user with a background image, login button, register link, and continue as guest button (for viewing gives annonymously)
+Render a background image, login button, register link, and continue as guest button (for viewing gives annonymously)
+
 
 //Guest
 Continue as Guest navigates to the Ask/Give page
@@ -18,19 +19,21 @@ Guests can view all asks and gives, as well as the resources page
 Selecting Register displays the registration page
 Selecting login displays the login page
 
-//Register
+//Register [Page: ]
 Register link navigates to a similar login page but with first name, last name, username (email) and password fields, (add location?) and Register button
 After registering, direct the user to the dashboard
 
-//Login
+//Login [Page: ]
 Login navigates to a login page with username (email) and password fields and Login button
 After logging in/registering, direct to the dashboard
 
-//Footer/Navigation
+//Home/Navigation [Page: ]
+
+//Footer [Page: ]
 If logged in, the footer contains dashboard link, ask/give link, resources link, and logout link
 If a guest, the footer contains register link, ask/give link, resources link, and login link
 
-//Dashboard
+//Dashboard [Page: ]
 Dashboard contains the logged-in user's name
 The create post form where new posts can be created
 Create post form contains fields for ask/give category, item type, title, description, item location (zip code)
@@ -41,13 +44,16 @@ Selecting edit post button loads the post into the edit post form where changes 
 Posts can be deleted (link next to post)
 Selecting delete post deletes the selected post
 
-//Ask_Give
+//Ask [Page: ]
 Top of page contains two buttons [Ask] and [Give] which act as filters. Either one or both can be selected.
 Page contains query/filter fields for type, location?
 Page content is a list of Posts
 Posts contain category, type, title, description, item location, and submitter's contact information
 
-//Resources
+//Give [Page: ]
+
+
+//Resources [Page: ]
 Hard coded list of local resources
 Hard coded list of national resources (on separate page?)
