+++ controllers/api/giveRoutes.js
@@ -1,33 +1,10 @@
 const router = require('express').Router();
 const { Ask_Give, User } = require('../../models');
-const withAuth = require('../../utils/auth')
+const withAuth = require('../../utils/auth');
 
 // Using the /gives endpoint
 
-//READ all gives
-router.get('/', withAuth, async (req, res) =>{
-    try{
-        const giveData = await Ask_Give.findAll({
-            where: {
-                ask_or_give: 'give',
-            },
-            attributes: ['title', 'content', 'resource_type', 'contact', 'createdAt', 'updatedAt'],
-            include: [
-                {
-                    model: User,
-                    attributes: ['name']
-                }
-            ]
-        });
-        const gives = giveData.map((ask_give) => ask_give.get({ plain: true }));
-        res.render('gives', {
-            gives,
-            loggedIn: req.session.logged_in
-        })
-    } catch (err) {
-        res.status(500).json(err);
-    }
-});
+
 
 //READ all gives by type of resource
 // router.get('/', withAuth, async (req, res) => {
