+++ controllers/api/userRoutes.js
@@ -1,7 +1,9 @@
 const router = require('express').Router();
-const { User } = require('../../models')
+const { User } = require('../../models');
 
 //Using the /user endpoint
+
+
 
 //Signup 
 router.post('/signup', async (req, res) => {
@@ -11,6 +13,7 @@ router.post('/signup', async (req, res) => {
         req.session.user_id = userData.id;
         req.session.logged_in = true;
         res.status(200).json(userData);
+        //need res.render navigation page when someone registers
       });
     } catch (err) {
       res.status(400).json(err);
@@ -40,6 +43,7 @@ router.post('/login', async (req, res) => {
         req.session.user_id = userData.id;
         req.session.logged_in = true;
         res.json({ user: userData, message: 'You are now logged in!' });
+        //need res.render navigation page when someone logs in
       });
     } catch (err) {
       res.status(400).json(err);
@@ -47,10 +51,12 @@ router.post('/login', async (req, res) => {
   });
 
   //Logout
+  //redirect user to the landing page
   router.post('/logout', (req, res) => {
     if (req.session.logged_in) {
       req.session.destroy(() => {
         res.status(204).end();
+        //present a snackbar/alert stating "You have logged out"
       });
     } else {
       res.status(404).end();
