+++ controllers/homeRoutes.js
@@ -1,11 +1,13 @@
 const router = require('express').Router();
-// const {Ask_Give} = require('../models');
+const { Ask_Give, User } = require('../models');
+const withAuth = require('../utils/auth');
+
 
 //Using the root endpoint
 
 //READ the homepage
 //works!
-router.get('/', async (req, res) => {
+router.get('/', (req, res) => {
     try {
 
         res.render ('homepage', {
@@ -52,10 +54,64 @@ router.get('/signup', (req, res) => {
 //     }
 //   });
 
+//READ all asks
+router.get('/asks', withAuth, async (req, res) =>{
+    try{
+        const askData = await Ask_Give.findAll({
+            where: {
+                ask_or_give: 'ask',
+            },
+            attributes: ['title', 'content', 'resource_type', 'contact', 'createdAt', 'updatedAt'],
+            include: [
+                {
+                    model: User,
+                    attributes: ['name']
+                }
+            ]
+        });
+        const asks = askData.map((ask_give) => ask_give.get({ plain: true }));
+        res.render('asks', {
+            asks,
+            loggedIn: req.session.logged_in
+        })
+    } catch (err) {
+        res.status(500).json(err);
+    }
+});
+
+//READ all gives
+router.get('/gives', withAuth, async (req, res) =>{
+    try{
+        const giveData = await Ask_Give.findAll({
+            where: {
+                ask_or_give: 'give',
+            },
+            attributes: ['title', 'content', 'resource_type', 'contact', 'createdAt', 'updatedAt'],
+            include: [
+                {
+                    model: User,
+                    attributes: ['name']
+                }
+            ]
+        });
+        const gives = giveData.map((ask_give) => ask_give.get({ plain: true }));
+        res.render('gives', {
+            gives,
+            loggedIn: req.session.logged_in
+        })
+    } catch (err) {
+        res.status(500).json(err);
+    }
+});
+
+//Need a GET route for the dashboard
+
+//Need a GET route for the navigation page
+
 //Resources page
 //works!
 router.get('/resources', (req, res) => {
-    res.render('national')
+    res.render('resources')
 });
 
 module.exports = router;
